+++ doc/whats-new.rst
@@ -40,6 +40,11 @@ Enhancements
  - Some speed improvement to construct :py:class:`~xarray.DataArrayRolling`
  object (:issue:`1993`)
  By `Keisuke Fujii <https://github.com/fujiisoup>`_.
+ - New option ``engine='h5netcdf-ng'`` for :py:meth:`~xarray.Dataset.to_netcdf`,
+   which uses the h5netcdf new API. This changes what settings are accepted
+   by the ``encoding`` parameter, and at the moment of writing is the only way
+   to use compression algorithms other than zlib, e.g. LZF (:ussue:`1536`).
+   By `Guido Imperiale <https://github.com/crusaderky>`_.
 
 Bug fixes
 ~~~~~~~~~